requirements.txt
asgiref==3.2.10
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorlog==4.2.1
constantly==15.1.0
cryptography==3.1
dateparser==0.7.6
decorator==4.4.2
Django==3.1.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-binance==0.7.5
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
urllib3==1.25.10
zope.interface==5.1.0
